(maint) Add tree alias
This commit adds an alias to give the linux program tree colors in
terminal. This is mostly for OSX. Might have to be reworked if used on
linux, because tree has colors by default on linx. (re: see ls alias
issue I had earlier on this repo between osx and linux)