Update custom ZSH theme
This commit updates my personal ZSH theme to be more similar to the
default Linux Mint bash prompt. It has ruby integration if you want, but
I left it commented out for now. In the future I'd like a zsh command to
toggle between it being displayed on the right and being gone.