Add mutt mail client automation
This commit adds some Mutt mail client automation for the dotfiles. It
will ask the user if they wish to use the mail client Mutt. If not, the
installer script will go on like normal and ignore any mutt related
things. If so, it will go ahead and ask the user for various config
settings to be set in the .muttrc.